new comment lines

+++ tgiann-anti-duplicate-vehicle/server.lua
@@ -1,4 +1,6 @@
-local alwaysCheck = false -- if set to false it will only be checked when the vehicle is created, If set to true all vehicles are checked every 5 seconds
+-- if set to false it will only be checked when the vehicle is created, If set to true all vehicles are checked every 5 seconds
+-- set this setting to true if NPC cars are enabled on your server (because every time a car is spawned, the entitycreating event will triggered and this will cause major performance issues.)
+local alwaysCheck = true
 
 if alwaysCheck then
 	CreateThread(function()
@@ -6,19 +8,21 @@ if alwaysCheck then
 			local plateList = {}
 			local allVeh = GetAllVehicles()
 			for i=1, #allVeh do
-				local plate = GetVehicleNumberPlateText(allVeh[i])
-				local model = GetEntityModel(allVeh[i])
-				if not plateList[plate] then 
-					plateList[plate] = model
-				elseif model == plateList[plate] then
-					DeleteEntity(allVeh[i])
+				if DoesEntityExist(allVeh[i]) then
+					local plate = GetVehicleNumberPlateText(allVeh[i])
+					local model = GetEntityModel(allVeh[i])
+					if not plateList[plate] then 
+						plateList[plate] = model
+					elseif model == plateList[plate] then
+						DeleteEntity(allVeh[i])
+					end
 				end
 			end
 			Wait(5000)
 		end
 	end)
 else
-	AddEventHandler("entityCreating", function(entity, data)
+	AddEventHandler("entityCreating", function(entity)
 		if DoesEntityExist(entity) and GetEntityType(entity) == 2 then
 			local myVehiclePlate = GetVehicleNumberPlateText(entity)
 			local myVehicleModel = GetEntityModel(entity)
